requirements-macos.txt
Flask
Werkzeug
keras
tensorflow-macos
Pillow
numpy
